Add more SKY130 cell libraries.
Also increase the documentation about skywater process.
